# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "ShapeShift",
      "description": "OpenAPI Schema for ShapeShift",
      "version": "0.0.1"
    },
    "paths": {
...
      "/api/v1/auth/login": {
        "post": {
          "tags": ["auth"],
          "summary": "Create access and refresh tokens for user",
          "operationId": "login",
          "requestBody": {
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/components/schemas/Body_login"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TokenSchema"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/auth/test-token": {
        "post": {
          "tags": ["auth"],
          "summary": "Test if the access token is valid",
          "operationId": "testToken",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserOut"
                  }
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      },
      "/api/v1/auth/refresh": {
        "post": {
          "tags": ["auth"],
          "summary": "Refresh token",
          "operationId": "refreshToken",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Refresh Token"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TokenSchema"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/exercise/": {
        "get": {
          "tags": ["exercises"],
          "summary": "Get all exercises",
          "operationId": "getAllExercises",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/ExerciseOut"
                    },
                    "type": "array",
                    "title": "Response Getallexercises"
                  }
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      },
      "/api/v1/exercise/exercises/{workout_id}": {
        "get": {
          "tags": ["exercises"],
          "summary": "Get all exercises",
          "operationId": "getExercises",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ExerciseOut"
                    },
                    "title": "Response Getexercises"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/exercise/{workout_id}": {
        "post": {
          "tags": ["exercises"],
          "summary": "Create an exercise",
          "operationId": "createExercise",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/exercise/{exercise_id}": {
        "get": {
          "tags": ["exercises"],
          "summary": "Get a single exercise",
          "operationId": "getExercise",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ExerciseOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["exercises"],
          "summary": "Update an exercise",
          "operationId": "updateExercise",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ExerciseOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["exercises"],
          "summary": "Delete an exercise",
          "operationId": "deleteExercise",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/exercise-instance/exercise-instances/exercise/{exercise_id}": {
        "get": {
          "tags": ["exercise-instances"],
          "summary": "Get all exercise instances by exercise",
          "operationId": "getExerciseInstancesByExercise",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ExerciseInstanceOut"
                    },
                    "title": "Response Getexerciseinstancesbyexercise"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Body_login": {
          "properties": {
            "grant_type": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "password"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grant Type"
            },
            "username": {
              "type": "string",
              "title": "Username"
            },
            "password": {
              "type": "string",
              "title": "Password"
            },
            "scope": {
              "type": "string",
              "title": "Scope",
              "default": ""
            },
            "client_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Id"
            },
            "client_secret": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Secret"
            }
          },
          "type": "object",
          "required": ["username", "password"],
          "title": "Body_login"
        },
...
        "Exercise": {
          "properties": {
            "_id": {
              "anyOf": [
                {
                  "type": "string",
                  "example": "5eb7cf5a86d9755df3a6c593"
                },
                {
                  "type": "null"
                }
              ],
              "title": " Id",
              "description": "MongoDB document ObjectID"
            },
            "exercise_id": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id",
              "unique": true
            },
            "name": {
              "type": "string",
              "maxLength": 50,
              "minLength": 1,
              "title": "Name"
            },
            "type": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExerciseType"
                },
                {
                  "type": "null"
                }
              ]
            },
            "equipment": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "muscle": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Muscle"
            },
            "track_reps": {
              "type": "boolean",
              "title": "Track Reps",
              "default": false
            },
            "track_weight": {
              "type": "boolean",
              "title": "Track Weight",
              "default": false
            },
            "track_distance": {
              "type": "boolean",
              "title": "Track Distance",
              "default": false
            },
            "track_duration": {
              "type": "boolean",
              "title": "Track Duration",
              "default": false
            },
            "workout": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Workout"
            },
            "owner": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Owner"
            }
          },
          "type": "object",
          "required": ["name", "workout", "owner"],
          "title": "Exercise"
        },
        "ExerciseCreate": {
          "properties": {
            "name": {
              "type": "string",
              "title": "Name"
            },
            "type": {
              "$ref": "#/components/schemas/ExerciseType"
            },
            "muscle": {
              "type": "string",
              "title": "Muscle"
            },
            "equipment": {
              "type": "string",
              "title": "Equipment"
            },
            "description": {
              "type": "string",
              "title": "Description"
            }
          },
          "type": "object",
          "required": ["name", "type", "muscle", "equipment", "description"],
          "title": "ExerciseCreate"
        },
...
        "ExerciseInstanceOut": {
          "properties": {
            "exercise_instance_id": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Instance Id"
            },
            "sets": {
              "type": "integer",
              "title": "Sets"
            },
            "reps": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Reps"
            },
            "weight": {
              "items": {
                "type": "number"
              },
              "type": "array",
              "title": "Weight"
            },
            "distance": {
              "items": {
                "type": "number"
              },
              "type": "array",
              "title": "Distance"
            },
            "duration": {
              "items": {
                "type": "string"
              },
              "type": "array",
              "title": "Duration"
            },
            "exercise": {
              "$ref": "#/components/schemas/Exercise"
            },
            "session": {
              "$ref": "#/components/schemas/Session"
            }
          },
          "type": "object",
          "required": [
            "exercise_instance_id",
            "sets",
            "reps",
            "weight",
            "distance",
            "duration",
            "exercise",
            "session"
          ],
          "title": "ExerciseInstanceOut"
        },
...
        "ExerciseOut": {
          "properties": {
            "exercise_id": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            },
            "name": {
              "type": "string",
              "title": "Name"
            },
            "type": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExerciseType"
                },
                {
                  "type": "null"
                }
              ]
            },
            "muscle": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Muscle"
            },
            "equipment": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "track_reps": {
              "type": "boolean",
              "title": "Track Reps"
            },
            "track_weight": {
              "type": "boolean",
              "title": "Track Weight"
            },
            "track_distance": {
              "type": "boolean",
              "title": "Track Distance"
            },
            "track_duration": {
              "type": "boolean",
              "title": "Track Duration"
            }
          },
          "type": "object",
          "required": [
            "exercise_id",
            "name",
            "type",
            "muscle",
            "equipment",
            "description",
            "track_reps",
            "track_weight",
            "track_distance",
            "track_duration"
          ],
          "title": "ExerciseOut"
        },
        "ExerciseType": {
          "type": "string",
          "enum": [
            "cardio",
            "strength",
            "olympic_weightlifting",
            "plyometrics",
            "powerlifting",
            "stretching",
            "strongman"
          ],
          "title": "ExerciseType"
        },
        "ExerciseUpdate": {
          "properties": {
            "name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            },
            "type": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExerciseType"
                },
                {
                  "type": "null"
                }
              ]
            },
            "muscle": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Muscle"
            },
            "equipment": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "track_reps": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Track Reps"
            },
            "track_weight": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Track Weight"
            },
            "track_distance": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Track Distance"
            },
            "track_duration": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Track Duration"
            }
          },
          "type": "object",
          "required": [
            "name",
            "type",
            "muscle",
            "equipment",
            "description",
            "track_reps",
            "track_weight",
            "track_distance",
            "track_duration"
          ],
          "title": "ExerciseUpdate"
        },
        "Gender": {
          "type": "string",
          "enum": ["male", "female", "other"],
          "title": "Gender"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
        "Session": {
          "properties": {
            "_id": {
              "anyOf": [
                {
                  "type": "string",
                  "example": "5eb7cf5a86d9755df3a6c593"
                },
                {
                  "type": "null"
                }
              ],
              "title": " Id",
              "description": "MongoDB document ObjectID"
            },
            "session_id": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id",
              "unique": true
            },
            "date": {
              "type": "string",
              "format": "date-time",
              "title": "Date"
            },
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "active": {
              "type": "boolean",
              "title": "Active",
              "default": false
            },
            "workout": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Workout"
            },
            "owner": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Owner"
            }
          },
          "type": "object",
          "required": ["workout", "owner"],
          "title": "Session"
        },
...
        "TokenSchema": {
          "properties": {
            "access_token": {
              "type": "string",
              "title": "Access Token"
            },
            "refresh_token": {
              "type": "string",
              "title": "Refresh Token"
            }
          },
          "type": "object",
          "required": ["access_token", "refresh_token"],
          "title": "TokenSchema"
        },
...
        "UserOut": {
          "properties": {
            "user_id": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            },
            "username": {
              "type": "string",
              "title": "Username"
            },
            "email": {
              "type": "string",
              "title": "Email"
            },
            "first_name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "First Name"
            },
            "last_name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last Name"
            },
            "birth_date": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Birth Date"
            },
            "gender": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Gender"
                },
                {
                  "type": "null"
                }
              ]
            },
            "height_in": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Height In"
            },
            "weight": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Weight"
            }
          },
          "type": "object",
          "required": [
            "user_id",
            "username",
            "email",
            "first_name",
            "last_name",
            "birth_date",
            "gender",
            "height_in",
            "weight"
          ],
          "title": "UserOut"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": ["loc", "msg", "type"],
          "title": "ValidationError"
        },
...
      },
      "securitySchemes": {
        "JWT": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {},
              "tokenUrl": "/api/v1/auth/login"
            }
          }
        }
      }
    },
    "servers": [
      {
        "url": "https://shapeshift-backend.onrender.com"
      }
    ]
  }